ref(widgets): Combine IssueWidgetCard into Widget Card
Also splits out the Context Menu render into its own
functional component since it was getting quite bloated.
Moved `transformTableResults` to issueWidgetQueries
so the component handles the data transform too.
