+++ static/app/views/dashboardsV2/widgetCard/widgetQueries.tsx
@@ -27,8 +27,8 @@ import {
 } from 'sentry/utils/discover/genericDiscoverQuery';
 import {TOP_N} from 'sentry/utils/discover/types';
 
-import {DisplayType, Widget, WidgetQuery} from './types';
-import {eventViewFromWidget} from './utils';
+import {DisplayType, Widget, WidgetQuery} from '../types';
+import {eventViewFromWidget} from '../utils';
 
 // Don't fetch more than 66 bins as we're plotting on a small area.
 const MAX_BIN_COUNT = 66;
