+++ tests/js/spec/views/dashboardsV2/widgetQueries.spec.jsx
@@ -2,7 +2,7 @@ import {mountWithTheme} from 'sentry-test/enzyme';
 import {initializeOrg} from 'sentry-test/initializeOrg';
 
 import {Client} from 'sentry/api';
-import WidgetQueries from 'sentry/views/dashboardsV2/widgetQueries';
+import WidgetQueries from 'sentry/views/dashboardsV2/widgetCard/widgetQueries';
 
 describe('Dashboards > WidgetQueries', function () {
   const initialData = initializeOrg({
